Create Hoo_bank
This is a react-js application built with tailwind CSS, CSS  and JavaScript. it is a plain Ui/Ux design with little reactivity but can be easily edited by a developer or none developer due to how reusable the components. This app main focus lies on the optimization of Code to produce a beautiful UI design.